+++ docs/src/components/LinkButton.tsx
@@ -3,18 +3,26 @@ import Link from 'next/link';
 import {ComponentProps} from 'react';
 
 type Props = {
+  className?: string;
   variant?: 'primary' | 'secondary';
+  size?: 'md' | 'lg';
 } & Omit<ComponentProps<typeof Link>, 'className'>;
 
 export default function LinkButton({
   children,
+  className,
+  size = 'md',
   variant = 'primary',
   ...rest
 }: Props) {
   return (
     <Link
       className={clsx(
-        'group inline-block rounded-full px-4 py-2 text-base font-semibold transition-colors lg:px-8 lg:py-4',
+        className,
+        'group inline-block font-semibold transition-colors',
+        size === 'lg'
+          ? 'rounded-full px-4 py-2 text-base lg:px-8 lg:py-4'
+          : 'rounded-lg px-3 py-1.5 text-sm',
         variant === 'primary'
           ? 'bg-slate-800 text-white hover:bg-slate-700 dark:bg-primary-400 dark:text-slate-900 dark:hover:bg-sky-200'
           : 'bg-slate-200 text-slate-700 dark:bg-slate-800 dark:text-white/90 dark:hover:bg-slate-700'
@@ -22,7 +30,14 @@ export default function LinkButton({
       {...rest}
     >
       {children}
-      <span className="ml-2 inline-block transition-transform group-hover:translate-x-1">
+      <span
+        className={clsx(
+          'ml-2 inline-block transition-transform',
+          size === 'lg'
+            ? 'group-hover:translate-x-1'
+            : 'group-hover:translate-x-0.5'
+        )}
+      >
         →
       </span>
     </Link>