+++ docs/src/pages/docs/workflows/_meta.tsx
@@ -2,7 +2,6 @@ export default {
   typescript: 'TypeScript augmentation',
   'localization-management': 'Localization management with Crowdin',
   'vscode-integration': 'VSCode integration',
-  linting: 'ESLint',
   storybook: 'Storybook',
   messages: 'Validating messages'
 };