+++ examples/example-app-router-playground/src/middleware.ts
@@ -1,7 +1,29 @@
-import createMiddleware from 'next-intl/middleware';
 import {routing} from './i18n/routing';
+import createMiddleware from 'next-intl/middleware';
+import {NextRequest, NextResponse} from 'next/server';
 
-export default createMiddleware(routing);
+export default async function middleware(request: NextRequest) {
+  const handleI18nRouting = createMiddleware(routing);
+  let response = handleI18nRouting(request);
+
+  // Additional rewrite when v2 cookie is set
+  if (response.ok) {
+    // (not for errors or redirects)
+    const [, locale, ...rest] = new URL(
+      response.headers.get('x-middleware-rewrite') || request.url
+    ).pathname.split('/');
+    const pathname = '/' + rest.join('/');
+
+    if (pathname === '/about' && request.cookies.get('v2')?.value === 'true') {
+      response = NextResponse.rewrite(
+        new URL(`/${locale}/about/v2`, request.url),
+        {headers: response.headers}
+      );
+    }
+  }
+
+  return response;
+}
 
 export const config = {
   matcher: [