+++ examples/example-app-router-playground/tests/main.spec.ts
@@ -469,15 +469,15 @@ it('can navigate between sibling pages that share a parent layout', async ({
 
 it('prefixes routes as necessary with the router', async ({page}) => {
   await page.goto('/');
-  page.getByTestId('ClientRouterWithoutProvider-link').click();
+  page.getByTestId('ClientRouter-link').click();
   await expect(page).toHaveURL('/nested');
 
   await page.goto('/en');
-  page.getByTestId('ClientRouterWithoutProvider-link').click();
+  page.getByTestId('ClientRouter-link').click();
   await expect(page).toHaveURL('/nested');
 
   await page.goto('/de');
-  page.getByTestId('ClientRouterWithoutProvider-link').click();
+  page.getByTestId('ClientRouter-link').click();
   await expect(page).toHaveURL('/de/verschachtelt');
 });
 
@@ -730,6 +730,26 @@ it('can switch the locale with `useRouter`', async ({page}) => {
   await expect(page).toHaveURL('/de/client');
   await page.getByRole('button', {name: 'Switch to en'}).click();
   await expect(page).toHaveURL('/client');
+});
+
+it('can use additional rewrites in the middleware', async ({browser}) => {
+  const context = await browser.newContext();
+  const page = await context.newPage();
+  await page.context().addCookies([
+    {
+      name: 'v2',
+      value: 'true',
+      url: 'http://localhost'
+    }
+  ]);
+
+  await page.goto('/about');
+  await page.getByRole('heading', {name: 'About v2'}).waitFor();
+
+  await page.goto('/de/about');
+  await page.getByRole('heading', {name: 'About v2'}).waitFor();
+
+  await context.close();
 });
 
 // https://github.com/radix-ui/primitives/issues/3165